protease/src/test/java/com/gitenter/protease/dao/git/CommitRepositoryTest.java
package com.gitenter.protease.dao.git;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

import java.io.IOException;

import org.eclipse.jgit.api.errors.GitAPIException;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ActiveProfiles;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.TestExecutionListeners;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
import org.springframework.test.context.support.DirtiesContextTestExecutionListener;
import org.springframework.test.context.transaction.TransactionalTestExecutionListener;

import com.gitenter.protease.ProteaseConfig;
import com.gitenter.protease.annotation.DbUnitMinimalDataSetup;
import com.github.springtestdbunit.DbUnitTestExecutionListener;
import com.github.springtestdbunit.annotation.DbUnitConfiguration;

@RunWith(SpringJUnit4ClassRunner.class)
@ActiveProfiles(profiles = "minimal")
@ContextConfiguration(classes=ProteaseConfig.class)
@TestExecutionListeners({
	DependencyInjectionTestExecutionListener.class,
	DirtiesContextTestExecutionListener.class,
	TransactionalTestExecutionListener.class,
	DbUnitTestExecutionListener.class })
@DbUnitConfiguration(databaseConnection={"schemaAuthDatabaseConnection", "schemaGitDatabaseConnection", "schemaReviewDatabaseConnection"})
public class CommitRepositoryTest {
	
	@Autowired CommitRepository repository;
	@Autowired DocumentRepository documentRepository;

	@Test
	@DbUnitMinimalDataSetup
	public void testDeleteById() throws IOException, GitAPIException {
		
		assertTrue(repository.findById(1).isPresent());
		assertTrue(documentRepository.findById(1).isPresent());
		
		repository.deleteById(1);
		
		/*
		 * Also confirmed that "ON DELETE CASCADE" is transfered to here.
		 */
		assertFalse(repository.findById(1).isPresent());
		assertFalse(documentRepository.findById(1).isPresent());
	}
}
